refactor(lang): move .unworklet/ seed routine from unplugin to lang
The `.unworklet/tsconfig.json` seed function was private to
packages/unplugin/src/index.ts, so unworklet-tsc (in packages/lang) has
no way to self-seed on a cold checkout. Pull the routine up to the
package the downstream dependency direction already points at
(unplugin → lang), so both unplugin and (in the next commit)
unworklet-tsc can call the same code.

Behavior change: zero. The unplugin runtime is byte-identical — same
function, imported from @unworklet/lang instead of defined locally.
No new runtime dependency in lang (node:fs + node:path only).

Setup for the next two commits (test + feat) that wire this into
unworklet-tsc.
